third_party/WebKit/LayoutTests/bluetooth/requestDevice.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharness-helpers.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="bluetooth-helpers.js"></script>
<script>
test(function(t) { assert_true(window.testRunner instanceof Object); t.done(); },
     "window.testRunner is required for the following tests.");

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('EmptyAdapter');
  return assert_promise_rejects(navigator.bluetooth.requestDevice(),
                                new TypeError());
}, 'Requires an argument.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('EmptyAdapter');
  return assert_promise_rejects(navigator.bluetooth.requestDevice({}),
                                new TypeError());
}, 'RequestDeviceOptions requires a |filters| member.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('EmptyAdapter');
  return navigator.bluetooth.requestDevice({
    filters: [{services: [genericAccessServiceUuid]}]
  }).then(function() {
    assert_unreached('Expected no devices but one was returned.');
  }, function(e) {
    assert_equals(e.name, 'NotFoundError');
      assert_equals(e.message, 'No Bluetooth devices in range.');
  });
}, 'Reject with NotFoundError.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('SingleEmptyDeviceAdapter');
  return navigator.bluetooth.requestDevice({
    filters: [{services: [genericAccessServiceUuid]}]
  }).then(function(device) {
    assert_equals(device.constructor.name, "BluetoothDevice");
  });
}, 'Mock will resolve.');
</script>
